make some *Controller interfaces for removing the @Mockable thing
