+++ app/src/commonMain/kotlin/de/westnordost/streetcomplete/data/download/strategy/MobileDataAutoDownloadStrategy.kt
@@ -1,13 +1,13 @@
 package de.westnordost.streetcomplete.data.download.strategy
 
 import de.westnordost.streetcomplete.data.download.tiles.DownloadedTilesSource
-import de.westnordost.streetcomplete.data.osm.mapdata.MapDataController
+import de.westnordost.streetcomplete.data.osm.mapdata.MapDataSource
 
 /** Download strategy if user is on mobile data */
 class MobileDataAutoDownloadStrategy(
-    mapDataController: MapDataController,
+    mapDataSource: MapDataSource,
     downloadedTilesSource: DownloadedTilesSource
-) : AVariableRadiusStrategy(mapDataController, downloadedTilesSource) {
+) : AVariableRadiusStrategy(mapDataSource, downloadedTilesSource) {
 
     override val maxDownloadAreaInKm2 = 10.0 // that's a radius of about 1.5 km
     override val desiredScoredMapDataCountInVicinity = 7500
