+++ app/src/commonMain/kotlin/de/westnordost/streetcomplete/data/download/strategy/WifiAutoDownloadStrategy.kt
@@ -1,13 +1,13 @@
 package de.westnordost.streetcomplete.data.download.strategy
 
 import de.westnordost.streetcomplete.data.download.tiles.DownloadedTilesSource
-import de.westnordost.streetcomplete.data.osm.mapdata.MapDataController
+import de.westnordost.streetcomplete.data.osm.mapdata.MapDataSource
 
 /** Download strategy if user is on wifi */
 class WifiAutoDownloadStrategy(
-    mapDataController: MapDataController,
+    mapDataSource: MapDataSource,
     downloadedTilesSource: DownloadedTilesSource
-) : AVariableRadiusStrategy(mapDataController, downloadedTilesSource) {
+) : AVariableRadiusStrategy(mapDataSource, downloadedTilesSource) {
 
     /** Let's assume that if the user is on wifi, he is either at home, at work, in the hotel, at a
      * café,... in any case, somewhere that would act as a "base" from which he can go on an
