+++ app/src/commonMain/kotlin/de/westnordost/streetcomplete/data/osm/created_elements/CreatedElementsModule.kt
@@ -6,5 +6,5 @@ val createdElementsModule = module {
     factory { CreatedElementsDao(get()) }
 
     single<CreatedElementsSource> { get<CreatedElementsController>() }
-    single { CreatedElementsController(get()) }
+    single<CreatedElementsController> { CreatedElementsControllerImpl(get()) }
 }
