+++ app/src/commonMain/kotlin/de/westnordost/streetcomplete/data/osm/mapdata/MapDataModule.kt
@@ -9,5 +9,6 @@ val mapDataModule = module {
     factory { RelationDao(get()) }
     factory { WayDao(get()) }
 
-    single { MapDataController(get(), get(), get(), get(), get(), get(), get()) }
+    single<MapDataSource> { get<MapDataController>() }
+    single<MapDataController> { MapDataControllerImpl(get(), get(), get(), get(), get(), get(), get()) }
 }
