+++ app/src/commonMain/kotlin/de/westnordost/streetcomplete/data/osm/mapdata/MapDataRepository.kt
@@ -1,5 +1,9 @@
 package de.westnordost.streetcomplete.data.osm.mapdata
 
+import de.westnordost.streetcomplete.data.osm.mapdata.ElementType.NODE
+import de.westnordost.streetcomplete.data.osm.mapdata.ElementType.RELATION
+import de.westnordost.streetcomplete.data.osm.mapdata.ElementType.WAY
+
 interface MapDataRepository {
     fun getNode(id: Long): Node?
     fun getWay(id: Long): Way?
@@ -13,9 +17,15 @@ interface MapDataRepository {
     fun getRelationsForWay(id: Long): Collection<Relation>
     fun getRelationsForRelation(id: Long): Collection<Relation>
 
-    fun get(type: ElementType, id: Long) = when (type) {
+    fun get(type: ElementType, id: Long): Element? = when (type) {
         ElementType.NODE     -> getNode(id)
         ElementType.WAY      -> getWay(id)
         ElementType.RELATION -> getRelation(id)
+    }
+
+    fun getRelationsForElement(type: ElementType, id: Long): Collection<Relation> = when (type) {
+        NODE -> getRelationsForNode(id)
+        WAY -> getRelationsForWay(id)
+        RELATION -> getRelationsForRelation(id)
     }
 }
