+++ app/src/commonMain/kotlin/de/westnordost/streetcomplete/data/osmnotes/NotesModule.kt
@@ -11,8 +11,9 @@ val notesModule = module {
     factory { NotesDownloader(get(), get()) }
     factory<PhotoServiceApiClient> { PhotoServiceApiClientImpl(get(), get(), ApplicationConstants.SC_PHOTO_SERVICE_URL) }
 
-    single {
-        NoteController(get()).apply {
+    single<NoteSource> { get<NoteController>() }
+    single<NoteController> {
+        NoteControllerImpl(get()).apply {
             // on notes have been updated, avatar images should be downloaded (cached) referenced in note discussions
             addListener(get<AvatarsInNotesUpdater>())
         }
