+++ app/src/commonMain/kotlin/de/westnordost/streetcomplete/data/osmnotes/edits/NoteEditsModule.kt
@@ -6,7 +6,7 @@ val noteEditsModule = module {
     factory { NoteEditsDao(get()) }
 
     single { NoteEditsUploader(get(), get(), get(), get(), get(), get(), get()) }
-    single { NoteEditsController(get()) }
+    single<NoteEditsController> { NoteEditsControllerImpl(get()) }
     single<NoteEditsSource> { get<NoteEditsController>() }
-    single { NotesWithEditsSource(get(), get(), get()) }
+    single<NotesWithEditsSource> { NotesWithEditsSourceImpl(get(), get(), get()) }
 }
