+++ app/src/commonMain/kotlin/de/westnordost/streetcomplete/data/visiblequests/HideQuestController.kt
@@ -3,5 +3,6 @@ package de.westnordost.streetcomplete.data.visiblequests
 import de.westnordost.streetcomplete.data.quest.QuestKey
 
 interface HideQuestController {
+    /** Mark the quest as hidden by user interaction */
     fun hide(key: QuestKey)
 }
