+++ app/src/commonMain/kotlin/de/westnordost/streetcomplete/data/visiblequests/VisibleQuestsModule.kt
@@ -12,7 +12,7 @@ val visibleQuestsModule = module {
     single { TeamModeQuestFilter(get(), get()) }
 
     single<QuestsHiddenSource> { get<QuestsHiddenController>() }
-    single { QuestsHiddenController(get(), get()) }
+    single<QuestsHiddenController> { QuestsHiddenControllerImpl(get(), get()) }
 
     single<VisibleEditTypeSource> { get<VisibleEditTypeController>() }
     single { VisibleEditTypeController(get(), get(), get()) }
